% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_of_hypothesis.R
\name{test_of_hypothesis}
\alias{test_of_hypothesis}
\title{Test of hypothesis which determines the p-value,degrees of freedom,confidence interval and test
statistic given the arguments}
\usage{
test_of_hypothesis(
  approach,
  independent = NULL,
  x,
  y = NULL,
  mu_0 = NULL,
  sigma_x = NULL,
  sigma_y = NULL,
  sigma_d = NULL,
  n1 = NULL,
  n2 = NULL,
  B = NULL,
  alpha,
  H1
)
}
\arguments{
\item{approach}{argument to determine whether the test is parametric or non-parametric}

\item{independent}{For matched pair data set, this parameter is false, and otherwise true.}

\item{x}{A random variable from a normal distribution}

\item{y}{Another random variable from a distribution}

\item{mu_0}{mean of a specific distribution}

\item{sigma_x}{population standard deviation of given data x}

\item{sigma_y}{population standard deviation of given data y}

\item{sigma_d}{population standard deviation of paired differences}

\item{n1}{number of rows matrix}

\item{n2}{number of columns matrix}

\item{B}{number of replication}

\item{alpha}{level of significance}

\item{H1}{alternate hypothesis}
}
\value{
p-value,degrees-of-freedom,confidence-interval,test-statistic
}
\description{
Test of hypothesis which determines the p-value,degrees of freedom,confidence interval and test
statistic given the arguments
}
\examples{
test_of_hypothesis(approach = "parametric", independent = FALSE, x = x,
y = y, mu_0 = mu_0, alpha = alpha, H1 = H1 )
}
